updates to backup_to_GNS() in .zshrc
- use --sparse file to force rsync to copy sparse files on the source
to sparse files at the destination.  I think this is why the backup
directory tree on raukawa was almost twice as large as the source tree
on my laptop: sparse files were being copied to
not-sparse (i.e. larger) files.  See:
https://serverfault.com/questions/655504/rsync-to-a-backup-server-resulted-in-additional-space-taken
- write logs to the laptop and also to raukawa.  Seems the laptop logs
do not capture deletions because they occur on the remote machine.
- add filters to protect /home/timh/.ssh and
/home/timh/.backaup_raukawa.log from deletion on the remote machine.
Deleting .ssh was removing my key for logging in without typing out
the password.